Extract ExecutorBase from executor.h Give GetExecutor external linkage.
