+++ dali/pipeline/executor/executor_test.cc
@@ -18,7 +18,7 @@
 #include "dali/core/tensor_shape.h"
 #include "dali/pipeline/data/backend.h"
 #include "dali/test/dali_test_decoder.h"
-#include "dali/pipeline/executor/executor.h"
+#include "dali/pipeline/executor/executor_impl.h"
 #include "dali/pipeline/executor/pipelined_executor.h"
 #include "dali/pipeline/executor/async_pipelined_executor.h"
 #include "dali/pipeline/executor/async_separated_pipelined_executor.h"
